chore: rename users to user, update route return status
