+++ server/src/api/v1/router.js
@@ -1,6 +1,6 @@
 import express from "express";
 import * as dotenv from "dotenv";
-import userRouter from "./users/user.route.js";
+import userRouter from "./user/user.route.js";
 import categoryRouter from "./categories/category.route.js";
 import sizeRouter from "./sizes/size.route.js";
 import colorRouter from "./colors/colors.route.js";
